refactor(platform): extract captured body sections and shared badge component
- Extract CapturedBodySections and related components to captured-body-sections.tsx
- Extract shared InlineBadge component to network-badge.tsx, replacing
  duplicated badge markup in TypeBadge, EncodingBadge, and TruncatedBadge
- Eliminate non-null assertions by narrowing via local variables
